feat: cache receipts
