+++ libs/web3/src/lib/use-transaction-receipt.ts
@@ -1,39 +1,57 @@
 import { useEffect, useState } from 'react';
-import { type TransactionReceipt } from '@ethersproject/providers';
+import type { TransactionReceipt } from '@ethersproject/providers';
 import { useDefaultWeb3Provider } from './default-web3-provider';
+import { useEthereumConfig } from './use-ethereum-config';
+
+const cache: Record<string, TransactionReceipt> = {};
 
 /**
- * Get and poll a transaction receipt until a certain number
- * of confirmations have been reached
+ * Get and cache a transaction receipt. Will poll until to
+ * get the latest number of confirmations until config.confirmations
+ * is reached
  */
 export const useTransactionReceipt = ({
   txHash,
   enabled = true,
-  confirmations = 1,
 }: {
   /* txHash of the transaction */
   txHash: string | null | undefined;
   /* enable or disable the request */
   enabled?: boolean;
-  /* number of confirmations at which point polling will stop */
-  confirmations?: number;
 }) => {
+  const { config } = useEthereumConfig();
   const { provider } = useDefaultWeb3Provider();
-  const [receipt, setReceipt] = useState<TransactionReceipt>();
+  const [receipt, setReceipt] = useState<TransactionReceipt | undefined>(() => {
+    if (!txHash) {
+      return;
+    }
+    return cache[txHash];
+  });
 
   useEffect(() => {
-    if (!provider || !txHash) return;
+    if (!config || !provider || !txHash) return;
 
     let mounted = true;
     let interval: NodeJS.Timer;
 
     const getReceipt = async () => {
+      const cachedReceipt = cache[txHash];
+
+      if (
+        cachedReceipt &&
+        cachedReceipt.confirmations >= config.confirmations
+      ) {
+        setReceipt(cachedReceipt);
+        return;
+      }
+
       const r = await provider.getTransactionReceipt(txHash);
 
       if (mounted) {
         setReceipt(r);
+        cache[txHash] = r;
 
-        if (r.confirmations >= confirmations) {
+        if (r.confirmations >= config.confirmations) {
           clearInterval(interval);
         }
       }
@@ -41,17 +59,14 @@ export const useTransactionReceipt = ({
 
     if (enabled) {
       getReceipt();
-
-      interval = setInterval(() => {
-        getReceipt();
-      }, 10000);
+      interval = setInterval(getReceipt, 10000);
     }
 
     return () => {
       mounted = false;
       clearInterval(interval);
     };
-  }, [provider, txHash, confirmations, enabled]);
+  }, [provider, txHash, config, enabled]);
 
   return { receipt };
 };
